[BinaryTranslation] changed type of address field to Long

+++ BinaryTranslation/deprecated/pt/up/fe/specs/binarytranslation/legacy/SimpleInstruction32.java
@@ -43,8 +43,8 @@ public interface SimpleInstruction32 extends Instruction {
     }
 
     @Override
-    default Integer getAddress() {
-        return getAddress32();
+    default Long getAddress() {
+        return Long.valueOf(getAddress32());
     }
 
     @Override
